# openapi.yaml
openapi: 3.0.3
info:
  title: Projeto Locadora de Filmes - Daniele Silva Santos
  description: API-REST destinada para controlar Filmes, Produtoras, Pessoas e entidades relacionadas
  version: 1.1.0
  contact:
    name: Daniele Silva Santos
    email: dev.daniele@outlook.com
  license:
    name: MIT
servers:
  - url: https://api-restful-locadora.onrender.com
    description: Ambiente de Produção
  - url: http://localhost:8080
    description: Ambiente Local
...
paths:
...
  /v1/senai/locadora/filme:
    get:
      tags: [FILME]
      summary: Listar todos os filmes
      responses:
        '200':
          description: Lista de filmes retornada com sucesso
        '404':
          description: Nenhum filme encontrado
        '500':
          description: Erro interno no servidor

    post:
      tags: [FILME]
      summary: Inserir novo filme
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilmeInput'
            example:
              nome: "Operação Tempestade de Aço"
              data_lancamento: "2025-03-20"
              duracao: "02:05:00"
              sinopse: "Um grupo de soldados de elite recebe uma missão impossível."
              avaliacao: "4.60"
              valor: "54.90"
              capa: "https://exemplo.com/capas/tempestade.jpg"
              trailer_url: "https://youtube.com/watch?v=temp003"
              status_filme: "Ativo"
              id_classificacao_indicativa: 3
              id_produtora: 3
              genero:
                - id: 2
                - id: 1
              idioma:
                - id: 1
                  tipo: "Original"
                - id: 2
                  tipo: "Dublado"
              pessoa:
                - id: 2
                  funcao: "Ator"
                - id: 1
                  funcao: "Dublador"
      responses:
        '201':
          description: Filme inserido com sucesso
        '400':
          description: Dados inválidos no body
        '415':
          description: Content-Type inválido (esperado application/json)
        '500':
          description: Erro interno no servidor
  /v1/senai/locadora/filme/{id}:
    get:
      tags: [FILME]
      summary: Buscar filme por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Filme encontrado
        '400':
          description: ID inválido
        '404':
          description: Filme não encontrado
        '500':
          description: Erro interno no servidor

    put:
      tags: [FILME]
      summary: Atualizar filme
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilmeInput'
      responses:
        '200':
          description: Filme atualizado com sucesso
        '400':
          description: Dados inválidos
        '404':
          description: Filme não encontrado
        '415':
          description: Content-Type inválido
        '500':
          description: Erro interno no servidor

    delete:
      tags: [FILME]
      summary: Excluir filme
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Filme excluído com sucesso
        '400':
          description: ID inválido
        '404':
          description: Filme não encontrado
        '500':
          description: Erro interno no servidor

# =========================================
#              CLASSIFICACAO
# =========================================
  /v1/senai/locadora/classificacao:
    get:
      tags: [CLASSIFICACAO]
      summary: Listar classificações
      responses:
        '200':
          description: Lista de classificações retornada com sucesso
        '404':
          description: Nenhuma classificação encontrada
        '500':
          description: Erro interno no servidor

    post:
      tags: [CLASSIFICACAO]
      summary: Inserir classificação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Classificacao'
            example:
              nome: "Livre"
              descricao: "Conteúdo adequado para todas as idades"
              sigla: "L"
              idade_minima: "0"
      responses:
        '201':
          description: Classificação criada com sucesso
        '400':
          description: Dados inválidos
        '415':
          description: Content-Type inválido
        '500':
          description: Erro interno no servidor
  /v1/senai/locadora/classificacao/{id}:
    get:
      tags: [CLASSIFICACAO]
      summary: Buscar classificação por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Classificação encontrada
        '400':
          description: ID inválido
        '404':
          description: Classificação não encontrada
        '500':
          description: Erro interno no servidor

    put:
      tags: [CLASSIFICACAO]
      summary: Atualizar classificação
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Classificacao'
      responses:
        '200':
          description: Classificação atualizada
        '400':
          description: Dados inválidos
        '404':
          description: Classificação não encontrada
        '415':
          description: Content-Type inválido
        '500':
          description: Erro interno no servidor

    delete:
      tags: [CLASSIFICACAO]
      summary: Excluir classificação
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Classificação excluída
        '400':
          description: ID inválido
        '404':
          description: Classificação não encontrada
        '500':
          description: Erro interno no servidor

# =========================================
#                  GENERO
# =========================================
  /v1/senai/locadora/genero:
    get:
      tags: [GENERO]
      summary: Listar gêneros
      responses:
        '200':
          description: Lista de gêneros retornada com sucesso
        '404':
          description: Nenhum gênero encontrado
        '500':
          description: Erro interno no servidor

    post:
      tags: [GENERO]
      summary: Inserir gênero
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Genero'
            example:
              tipo: "Ação"
              descricao: "Filmes com muita adrenalina e combates"
      responses:
        '201':
          description: Gênero criado com sucesso
        '400':
          description: Dados inválidos
        '415':
          description: Content-Type inválido
        '500':
          description: Erro interno no servidor
  /v1/senai/locadora/genero/{id}:
    get:
      tags: [GENERO]
      summary: Buscar gênero por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Gênero encontrado
        '400':
          description: ID inválido
        '404':
          description: Gênero não encontrado
        '500':
          description: Erro interno no servidor

    put:
      tags: [GENERO]
      summary: Atualizar gênero
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Genero'
      responses:
        '200':
          description: Gênero atualizado
        '400':
          description: Dados inválidos
        '404':
          description: Gênero não encontrado
        '415':
          description: Content-Type inválido
        '500':
          description: Erro interno no servidor

    delete:
      tags: [GENERO]
      summary: Excluir gênero
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Gênero excluído
        '400':
          description: ID inválido
        '404':
          description: Gênero não encontrado
        '500':
          description: Erro interno no servidor

# =========================================
#                  IDIOMA
# =========================================
  /v1/senai/locadora/idioma:
    get:
      tags: [IDIOMA]
      summary: Listar idiomas
      responses:
        '200':
          description: Lista de idiomas retornada com sucesso
        '404':
          description: Nenhum idioma encontrado
        '500':
          description: Erro interno no servidor

    post:
      tags: [IDIOMA]
      summary: Inserir idioma
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Idioma'
            example:
              nome: "Português"
              sigla: "PT-BR"
      responses:
        '201':
          description: Idioma criado com sucesso
        '400':
          description: Dados inválidos
        '415':
          description: Content-Type inválido
        '500':
          description: Erro interno no servidor
  /v1/senai/locadora/idioma/{id}:
    get:
      tags: [IDIOMA]
      summary: Buscar idioma por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Idioma encontrado
        '400':
          description: ID inválido
        '404':
          description: Idioma não encontrado
        '500':
          description: Erro interno no servidor

    put:
      tags: [IDIOMA]
      summary: Atualizar idioma
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Idioma'
      responses:
        '200':
          description: Idioma atualizado
        '400':
          description: Dados inválidos
        '404':
          description: Idioma não encontrado
        '415':
          description: Content-Type inválido
        '500':
          description: Erro interno no servidor

    delete:
      tags: [IDIOMA]
      summary: Excluir idioma
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Idioma excluído
        '400':
          description: ID inválido
        '404':
          description: Idioma não encontrado
        '500':
          description: Erro interno no servidor

# =========================================
#               NACIONALIDADE
# =========================================
  /v1/senai/locadora/nacionalidade:
    get:
      tags: [NACIONALIDADE]
      summary: Listar nacionalidades
      responses:
        '200':
          description: Lista de nacionalidades retornada com sucesso
        '404':
          description: Nenhuma nacionalidade encontrada
        '500':
          description: Erro interno no servidor

    post:
      tags: [NACIONALIDADE]
      summary: Inserir nacionalidade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Nacionalidade'
            example:
              nome_pais: "Brasil"
      responses:
        '201':
          description: Nacionalidade criada com sucesso
        '400':
          description: Dados inválidos
        '415':
          description: Content-Type inválido
        '500':
          description: Erro interno no servidor
  /v1/senai/locadora/nacionalidade/{id}:
    get:
      tags: [NACIONALIDADE]
      summary: Buscar nacionalidade por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Nacionalidade encontrada
        '400':
          description: ID inválido
        '404':
          description: Nacionalidade não encontrada
        '500':
          description: Erro interno no servidor

    put:
      tags: [NACIONALIDADE]
      summary: Atualizar nacionalidade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Nacionalidade'
      responses:
        '200':
          description: Nacionalidade atualizada
        '400':
          description: Dados inválidos
        '404':
          description: Nacionalidade não encontrada
        '415':
          description: Content-Type inválido
        '500':
          description: Erro interno no servidor

    delete:
      tags: [NACIONALIDADE]
      summary: Excluir nacionalidade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Nacionalidade excluída
        '400':
          description: ID inválido
        '404':
          description: Nacionalidade não encontrada
        '500':
          description: Erro interno no servidor

# =========================================
#                   SEXO
# =========================================
  /v1/senai/locadora/sexo:
    get:
      tags: [SEXO]
      summary: Listar sexos
      responses:
        '200':
          description: Lista retornada com sucesso
        '404':
          description: Nenhum registro encontrado
        '500':
          description: Erro interno no servidor

    post:
      tags: [SEXO]
      summary: Inserir sexo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sexo'
            example:
              nome: "Masculino"
              sigla: "M"
      responses:
        '201':
          description: Criado com sucesso
        '400':
          description: Dados inválidos
        '415':
          description: Content-Type inválido
        '500':
          description: Erro interno no servidor
...
  /v1/senai/locadora/tipo-telefone:
    get:
      tags: [TIPO_TELEFONE]
      summary: Listar tipos de telefone
      responses:
        '200':
          description: Lista retornada com sucesso
        '404':
          description: Nenhum registro encontrado
        '500':
          description: Erro interno no servidor

    post:
      tags: [TIPO_TELEFONE]
      summary: Inserir tipo de telefone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TipoTelefone'
            example:
              tipo: "Celular"
      responses:
        '201':
          description: Criado com sucesso
        '400':
          description: Dados inválidos
        '415':
          description: Content-Type inválido
        '500':
          description: Erro interno no servidor
...
components:
  schemas:
...
    Classificacao:
      type: object
      required: [nome, descricao, sigla, idade_minima]
      properties:
        nome:
          type: string
          example: "Livre"
        descricao:
          type: string
          example: "Conteúdo adequado para todas as idades"
        sigla:
          type: string
          example: "L"
        idade_minima:
          type: integer
          example: 0
    Genero:
      type: object
      required: [tipo, descricao]
      properties:
        tipo:
          type: string
          maxLength: 30
          example: "Ação"
        descricao:
          type: string
          example: "Filmes com muita adrenalina e combates"
    Idioma:
      type: object
      required: [nome, sigla]
      properties:
        nome:
          type: string
          maxLength: 100
          example: "Português"
        sigla:
          type: string
          maxLength: 8
          example: "PT-BR"
    Sexo:
      type: object
      required: [nome, sigla]
      properties:
        nome:
          type: string
          example: "Masculino"
        sigla:
          type: string
          example: "M"
    Nacionalidade:
      type: object
      required: [nome_pais]
      properties:
        nome_pais:
          type: string
          example: "Brasil"
    TipoTelefone:
      type: object
      required: [tipo]
      properties:
        tipo:
          type: string
          example: "Celular"
...